#42 Move type-interfaces to g2api
