+++ g2diagnostic/main.go
@@ -1,46 +1,5 @@
 package g2diagnostic
 
-import (
-	"context"
-
-	"github.com/senzing/go-logging/logger"
-	"github.com/senzing/go-observing/observer"
-)
-
-// ----------------------------------------------------------------------------
-// Types
-// ----------------------------------------------------------------------------
-
-// The G2diagnostic interface is a Golang representation of Senzing's libg2diagnostic.h
-type G2diagnostic interface {
-	CheckDBPerf(ctx context.Context, secondsToRun int) (string, error)
-	CloseEntityListBySize(ctx context.Context, entityListBySizeHandle uintptr) error
-	Destroy(ctx context.Context) error
-	FetchNextEntityBySize(ctx context.Context, entityListBySizeHandle uintptr) (string, error)
-	FindEntitiesByFeatureIDs(ctx context.Context, features string) (string, error)
-	GetAvailableMemory(ctx context.Context) (int64, error)
-	GetDataSourceCounts(ctx context.Context) (string, error)
-	GetDBInfo(ctx context.Context) (string, error)
-	GetEntityDetails(ctx context.Context, entityID int64, includeInternalFeatures int) (string, error)
-	GetEntityListBySize(ctx context.Context, entitySize int) (uintptr, error)
-	GetEntityResume(ctx context.Context, entityID int64) (string, error)
-	GetEntitySizeBreakdown(ctx context.Context, minimumEntitySize int, includeInternalFeatures int) (string, error)
-	GetFeature(ctx context.Context, libFeatID int64) (string, error)
-	GetGenericFeatures(ctx context.Context, featureType string, maximumEstimatedCount int) (string, error)
-	GetLogicalCores(ctx context.Context) (int, error)
-	GetMappingStatistics(ctx context.Context, includeInternalFeatures int) (string, error)
-	GetPhysicalCores(ctx context.Context) (int, error)
-	GetRelationshipDetails(ctx context.Context, relationshipID int64, includeInternalFeatures int) (string, error)
-	GetResolutionStatistics(ctx context.Context) (string, error)
-	GetTotalSystemMemory(ctx context.Context) (int64, error)
-	Init(ctx context.Context, moduleName string, iniParams string, verboseLogging int) error
-	InitWithConfigID(ctx context.Context, moduleName string, iniParams string, initConfigID int64, verboseLogging int) error
-	RegisterObserver(ctx context.Context, observer observer.Observer) error
-	Reinit(ctx context.Context, initConfigID int64) error
-	SetLogLevel(ctx context.Context, logLevel logger.Level) error
-	UnregisterObserver(ctx context.Context, observer observer.Observer) error
-}
-
 // ----------------------------------------------------------------------------
 // Variables
 // ----------------------------------------------------------------------------
